refactor: verify production numeric kernels in place, drop Stainless mirror
Move the analysis engine's collection-free numeric/geometric kernels into a
new PureKernels object and point `stainlessVerify` at it directly, so the
formally-verified code IS the code that runs — no separate mirror to drift.

- Add PureKernels (instability, pageRankBase, rangeContains, rangeSpan,
  nextLevel) as the single Stainless verification target.
- nextLevel: new verified function — the longest-path layering recurrence
  (monotonic, >=0, overflow-guarded), extracted from GraphMetrics.layers.
- GraphMetrics.instability/pageRankBase delegate to PureKernels; layers uses
  nextLevel instead of inline `math.max(acc, sv + 1)`.
- AnalyzerHelpers.rangeContains/rangeSpan become thin s.Range adapters over
  PureKernels, decoupling the range geometry from the SemanticDB protobuf.
- Delete StainlessContracts.scala (the collection-free mirror); retarget
  scripts/stainless-verify.sh and the stainlessVerify task to PureKernels.
- Bundle stainless-library.jar into the mcp fat jar: PureKernels imports
  stainless.lang.* so pageRankBase's NaN guard can use stainless's IEEE-754
  Double model (.isNaN), the only not-NaN witness the verifier accepts.

Gate: 29 valid, 0 invalid (rangeSpan's nonlinear multiplication VCs tolerated
as unknown under bundled smt-z3, valid under native Z3). All analysis + mcp
suites pass; shipped fat jar smoke-tested (structure tool runs PureKernels at
runtime with no NoClassDefFoundError).
