+++ analysis/src/main/scala/com/github/mercurievv/scalasemantic/analysis/graph/GraphMetrics.scala
@@ -1,5 +1,6 @@
 package com.github.mercurievv.scalasemantic.analysis.graph
 
+import com.github.mercurievv.scalasemantic.analysis.PureKernels
 import stainless.annotation.opaque
 import stainless.annotation.pure
 
@@ -8,12 +9,14 @@ import stainless.annotation.pure
   * wartremover rules; recursion depth is bounded by the graph's depth, fine for project-sized
   * graphs.
   *
-  * Stainless formal-verification targets (`sbt stainlessVerify` runs the standalone tool from
-  * https://github.com/epfl-lara/stainless/releases/tag/v0.9.9.3 over the collection-free mirrors in
-  * [[com.github.mercurievv.scalasemantic.analysis.StainlessContracts]]):
-  *   - [[instability]]: `@pure` + domain/overflow require + `[0,1]` and boundary-pinning ensuring
-  *   - [[pageRankBase]]: `@pure` — teleport probability proven ∈ [0,1] (NaN-guarded)
-  *   - [[reverse]]: `@pure @opaque` — pure, body treated as a black box by callers
+  * The collection-free numeric kernels these algorithms rely on
+  * ([[com.github.mercurievv.scalasemantic.analysis.PureKernels.instability instability]],
+  * [[com.github.mercurievv.scalasemantic.analysis.PureKernels.pageRankBase pageRankBase]],
+  * [[com.github.mercurievv.scalasemantic.analysis.PureKernels.nextLevel nextLevel]]) live in
+  * [[com.github.mercurievv.scalasemantic.analysis.PureKernels]] and are formally verified there by
+  * `sbt stainlessVerify` — the standalone Stainless tool can't extract this file (it doesn't model
+  * `.iterator`/`.view`/`.groupBy`), so the verifiable logic is delegated rather than mirrored.
+  * [[reverse]] is `@pure @opaque`: pure, its body treated as a black box by callers.
   */
 object GraphMetrics:
 
@@ -37,39 +40,19 @@ object GraphMetrics:
       .toMap
       .ensuring(res => res.keySet == nodes)
 
-  /** Instability Ce/(Ca+Ce), or 0 for an isolated node (no edges either way).
-    *
-    * Stainless contract (formally verified — see [[StainlessContracts.instability]]):
-    *   - Precondition: afferent and efferent counts are non-negative and `ca + ce` does not
-    *     overflow `Int` (the `ca <= Int.MaxValue - ce` bound — Stainless found that overflow in
-    *     `ca + ce` invalidates the postcondition otherwise).
-    *   - Postcondition: result ∈ [0.0, 1.0], AND the boundaries are pinned exactly — no out-edges ⟹
-    *     0.0 (maximally stable), only out-edges ⟹ 1.0 (maximally unstable). Pinning the boundaries
-    *     (not just the range) is what catches an operand swap in the formula. The two extra clauses
-    *     are `A ==> B` written as `!A || B` to avoid importing `stainless.lang` into production
-    *     code.
+  /** Instability Ce/(Ca+Ce), or 0 for an isolated node — delegates to the formally-verified
+    * [[com.github.mercurievv.scalasemantic.analysis.PureKernels.instability]] (result ∈ [0,1] with
+    * boundaries pinned; overflow-guarded).
     */
   @pure
-  def instability(ca: Int, ce: Int): Double =
-    require(ca >= 0 && ce >= 0 && ca <= Int.MaxValue - ce)
-    (if ca + ce == 0 then 0.0 else ce.toDouble / (ca + ce)).ensuring(r =>
-      r >= 0.0 && r <= 1.0 &&
-        (ce != 0 || r == 0.0) && // ce == 0  ==> r == 0.0
-        (ca != 0 || ce == 0 || r == 1.0) // ca == 0 && ce > 0  ==> r == 1.0
-    )
+  def instability(ca: Int, ce: Int): Double = PureKernels.instability(ca, ce)
 
-  /** PageRank teleport term `(1 - damping) / n`: the rank mass every node receives each iteration
-    * regardless of the link structure. It is a probability and must lie in [0.0, 1.0].
-    *
-    * Stainless contract (formally verified — see [[StainlessContracts.pageRankBase]]):
-    *   - Precondition: `damping` is a probability in [0, 1] and not NaN (Stainless reports a
-    *     `Comparison with NaN` counter-example without the `!isNaN` guard), and `n > 0`.
-    *   - Postcondition: result ∈ [0.0, 1.0].
+  /** PageRank teleport term `(1 - damping) / n` — delegates to the formally-verified
+    * [[com.github.mercurievv.scalasemantic.analysis.PureKernels.pageRankBase]] (probability ∈
+    * [0,1], NaN-guarded).
     */
   @pure
-  def pageRankBase(damping: Double, n: Int): Double =
-    require(!damping.isNaN && damping >= 0.0 && damping <= 1.0 && n > 0)
-    ((1.0 - damping) / n).ensuring(r => r >= 0.0 && r <= 1.0)
+  def pageRankBase(damping: Double, n: Int): Double = PureKernels.pageRankBase(damping, n)
 
   /** Strongly-connected components (Kosaraju): a list of node sets. Singleton sets are acyclic
     * nodes; a set of size > 1 is a dependency cycle (mutually-reachable nodes).
@@ -117,7 +100,7 @@ object GraphMetrics:
           val (lv, m) =
             condensed.getOrElse(c, Set.empty).foldLeft((0, memo)) { case ((acc, mm), s) =>
               val (sv, mm2) = level(s, mm)
-              (math.max(acc, sv + 1), mm2)
+              (PureKernels.nextLevel(acc, sv), mm2)
             }
           (lv, m + (c -> lv))
     val componentLevels = components.indices.foldLeft(Map.empty[Int, Int]) { (memo, c) =>
