+++ build.sbt
@@ -170,18 +170,21 @@ lazy val pc = (project in file("pc"))
 // Stainless library: `stainless-library_3` is NOT published to Maven Central, so we DON'T depend on
 // it as a managed artifact (that would fail to resolve in a clean CI checkout). Instead the v0.9.9.3
 // release jar is checked in at `analysis/lib/stainless-library.jar` and picked up as an UNMANAGED
-// dependency (sbt auto-includes jars under a module's `lib/`). It is compile-only in spirit — it
-// supplies @pure/@opaque and stainless.lang for the contracts — and is explicitly excluded from the
-// mcp fat jar below so it never ships at runtime. The sbt-stainless plugin from the same release is
-// sbt-1.x-only and can't load under sbt 2.0, so verification runs via the standalone tool
-// (`sbt stainlessVerify`), not at compile time.
+// dependency (sbt auto-includes jars under a module's `lib/`). It supplies @pure/@opaque and
+// stainless.lang for the contracts. It IS bundled into the mcp fat jar (see the assembly settings
+// below): `PureKernels` imports `stainless.lang.*` for the IEEE-754 `Double` model its NaN guard
+// needs, leaving a runtime reference; under that import `require`/`ensuring` are erased ghosts, so
+// the bundled jar costs classpath presence but no per-call overhead. The sbt-stainless plugin from
+// the same release is sbt-1.x-only and can't load under sbt 2.0, so verification runs via the
+// standalone tool (`sbt stainlessVerify`), not at compile time.
 
-// Formal-verification gate: run the standalone Stainless tool over StainlessContracts.scala via
+// Formal-verification gate: run the standalone Stainless tool over PureKernels.scala (the
+// production numeric/geometric kernels, verified in place — no mirror) via
 // scripts/stainless-verify.sh (which downloads + caches the tool, bundled solvers included). The
 // script fails iff a VC is INVALID, tolerating the `unknown`/timeout VCs the bundled smt-z3 cannot
 // discharge. CI runs `sbt stainlessVerify`; devs can too.
 lazy val stainlessVerify =
-  taskKey[Unit]("Formally verify StainlessContracts.scala with the standalone Stainless tool")
+  taskKey[Unit]("Formally verify PureKernels.scala with the standalone Stainless tool")
 
 lazy val analysis = (project in file("analysis"))
   .dependsOn(core, pc)
@@ -221,16 +224,12 @@ lazy val mcp = (project in file("mcp"))
     // to GitHub Releases. Pin the main class (the module has two @main) so the manifest is correct.
     assembly / mainClass := Some("com.github.mercurievv.scalasemantic.mcpServer"),
     assembly / assemblyJarName := "scalasemantic-mcp.jar",
-    // stainless-library reaches this classpath as an unmanaged jar from analysis/lib/ (it backs the
-    // compile-time @pure/@opaque contracts only). Keep it out of the runtime fat jar — the
-    // annotations are never resolved at runtime and the contracts are verified offline.
-    assembly / assemblyExcludedJars := {
-      // sbt 2.0 classpath entries are virtual-file refs; resolve to real paths to match by name.
-      val converter = fileConverter.value
-      (assembly / fullClasspath).value.filter { af =>
-        converter.toPath(af.data).getFileName.toString == "stainless-library.jar"
-      }
-    },
+    // stainless-library reaches this classpath as an unmanaged jar from analysis/lib/. It IS bundled
+    // into the fat jar: `PureKernels` imports `stainless.lang.*` so its `pageRankBase` contract can
+    // use stainless's IEEE-754 `Double` model (`.isNaN`) — the only not-NaN witness the verifier
+    // accepts — which leaves a runtime reference to the stainless namespace. Under the import,
+    // `require`/`ensuring` become stainless's erased ghost variants (zero runtime cost), so shipping
+    // the jar adds the dependency on the classpath without adding per-call checking overhead.
     assembly / assemblyMergeStrategy := {
       case PathList("META-INF", "MANIFEST.MF") => MergeStrategy.discard
       case PathList("META-INF", xs @ _*)
